pyproject.toml
[tool.poetry]
name = "auto-retail-aic2022"
version = "0.1.0"
description = ""
authors = ["duytue <tue@vietkite.com>"]

[tool.poetry.dependencies]
python = "^3.8.5"
pytorch-lightning = "^1.5.10"
numpy = "^1.22.3"
python-box = "^5.4.1"
opencv-python = "^4.5.5"
pandas = "^1.4.1"
torchvision = "^0.12.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
black = "^22.1.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
